% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.Aph.R
\name{compute.Aph}
\alias{compute.Aph}
\title{Compute phytoplankton absorption spectrum (Aph) from total particulate absorption (Ap)
and, if available, the non-algal particles absorption (Anap) spectra}
\usage{
compute.Aph(waves, Ap, Anap)
}
\arguments{
\item{waves}{is a vector of wavelengths}

\item{Ap}{is the total particulate absorption vector (mandatory)}

\item{Anap}{is the non-algal particles absorption vector. If Anap=NA
only the spectral deconvolution method of Bricaud and Stramski (1990) is applied.}
}
\value{
It returns a list with all the cumputed spectra of Aph and Anap
as well as the NAP spectral slopes for the different method. The program computes the Aph using three approaches:
\itemize{
   \item{1. The Anap is subtracted from Ap after adjusting the NIR of Anap
such as Ap(NIR)==Anap(NIR) to obtain Aph.offset;}
   \item{2. An exponential fit is applied to the Anap spectrum (Anap.fitted) avoiding
the blue and red portions of the spectrum where some remaining
pigments on the filter after methanol extraction may be present.}
   \item{3. A spectral deconvolution following the method of Bricaud and Stramski (1990)
is performed. Two equations are applied to Ap spectra (their eq. 5 and eq 6) yielding
Aph.BS90, Anap.BS90 and Aph.BS90.v2, Anap.BS90.v2.}
}
}
\description{
Compute phytoplankton absorption spectrum (Aph) from total particulate absorption (Ap)
and, if available, the non-algal particles absorption (Anap) spectra
}
\author{
Simon Bélanger
}
