% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.discrete.ATOT.for.COPS.R
\name{compute.discrete.aTOT.for.COPS}
\alias{compute.discrete.aTOT.for.COPS}
\title{Compute the total absorption for COPS processing of LuZ profile}
\usage{
compute.discrete.aTOT.for.COPS(cops.path = "./", ag.RData = "ag.RData",
  ap.RData = "ap.RData")
}
\arguments{
\item{cops.path}{is the full path of the COPS data where the absorption.cops.dat file can be found}

\item{ag.RData}{is the full path and file name of the RData file contaning the CDOM absorption spectrum}

\item{ap.RData}{is the full path and file name of the RData file contaning the particle absorption spectrum}
}
\value{
It returns a vector of total absorption.
}
\description{
Compute the total absorption for COPS instrument self-shadow
correction of LuZ. It is run only after a preliminary processing
of the spectrophotometric data using the \pkg{Rspectro} package.
}
\details{
The program will open a RData file located in ../COPS/BIN folder to get
the wavelengths available on the COPS. It will edit the file
../COPS/absorption.cops.dat and produce a PNG figure showing the spectral
absorption that will be used in the COPS processing for instrument self-shadow
correction.

The pure water absorption is taken from a composite table
(Kou etal 1993, Pope and Fry 1997, etc).
}
\author{
Simon Belanger
}
