% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.Ag.DB.R
\name{generate.Ag.DB}
\alias{generate.Ag.DB}
\title{Assemble the CDOM absorption data into a single data base}
\usage{
generate.Ag.DB(
  log.file = "Ag_log_TEMPLATE.dat",
  data.path = "./",
  MISSION = "YYY",
  MEASURED = FALSE
)
}
\arguments{
\item{log.file}{is the name of the ASCII file containing the
list of ID to process (see details below).}

\item{data.path}{is the path where the RData folder is located.
So the files must be stored in data.path/RData.  Default is "./".}

\item{MISSION}{is a character string that will be used to name
the output files. The default is "YYY".
(i.e. MISSION.Ag.RData; MISSION.Ag.dat;
MISSION.Fitted.params.dat; MISSION.Ag.png)}

\item{MEASURED}{if a logical parameter indicating if the fitted parameter is computed using the measured Ag without the red offset. Defautl is FALSE}
}
\value{
the function returns a list containing the following fields:
  waves, ID, Station, Depth,Ag.raw, Ag.offset,S275_295,
  S350_400, S350_500, Sr, K, a440.

  This list (Ag.DB) is save in MISSION.Ag.RData.
  The Ag.offset spectra are saved in ASCII format in MISSION.Ag.dat,
  while the fitting parameters are saved in MISSION.Fitted.params.dat.
  Finally, a PNG file showing all CDOM spectra ise created.
}
\description{
This function assembles the CDOM data in a single data base and
generate 1) one RData file, 2) two ASCII files and 3) one plot
presenting all CDOM spectra.
}
\seealso{
\code{\link{process.Ag}}, \code{\link{run.process.Ag.batch}}
}
\author{
Simon Bélanger
}
