% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.Ap.DB.R
\name{generate.Ap.DB}
\alias{generate.Ap.DB}
\title{Assemble the Ap absorption data into a single data base}
\usage{
generate.Ap.DB(
  log.file = "Ap_log_TEMPLATE.dat",
  data.path = "./",
  MISSION = "YYY",
  Beta = "Stramski"
)
}
\arguments{
\item{log.file}{is the name of the ASCII file containing the
list of ID to process (see details below).}

\item{data.path}{is the path where the RData folder is located.
So the files must be stored in data.path/RData.  Default is "./".}

\item{MISSION}{is a character string that will be used to name
the output files. The default is "YYY".
(i.e. MISSION.Ap.RData; MISSION.Ap.dat;
MISSION.Fitted.params.dat; MISSION.Ap.png)}

\item{Beta}{if the amplification factor your want to select. Default is Stramski.}
}
\value{

}
\description{
This function assembles the Ap data in a single data base and
generate 1) three RData file, 2) three ASCII files and 3) three plot
presenting all Ap, Anap and Aphy spectra.
}
\seealso{
\code{\link{process.Ap}}, \code{\link{run.process.Ap.batch}}
}
\author{
Simon Bélanger
}
