% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Ag.R
\name{plot.Ag}
\alias{plot.Ag}
\title{Plot CDOM spectra}
\usage{
\method{plot}{Ag}(path.png, Ag, PNG = TRUE)
}
\arguments{
\item{path.png}{is the path where the png file will be saved}

\item{Ag}{is a list return by \code{\link{process.Ag}}.}

\item{PNG}{is a logical parameter indicating whether or not
the plots are save in a PNG file. Default is TRUE.}
}
\description{
Plot the CDOM spectra stored in an Ag list returned by
\code{\link{process.Ag}}.
}
\details{
Four plots are produced with
the same three spectra but with different z-axis and x-axis.
The three spectra are the raw (black), the corrected spectra
assuming NULL point at 690 nm (red) and the fitted spectra
on the 350 to 500 nm range (green).
}
\examples{
setwd(path.package("RspectroAbs"))
sample = read.LAMBDA850("data/433423.Sample.Raw.csv1")
Ag = process.Ag(sample, ID=433423,
               Station="L3_20.5", Depth=99,pathlength=0.1)
plot.Ag("./",Ag, PNG=FALSE)

}
\seealso{
\code{\link{process.Ag}}
}
\author{
Simon Belanger
}
