% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.Ag.R
\name{process.Ag}
\alias{process.Ag}
\title{Process raw optical depth to CDOM spectra and analyse the spectral
 shape using \pkg{cdom} package functionalities.}
\usage{
process.Ag(sample, ID, Station, Depth, pathlength, DilutionFactor = 1)
}
\arguments{
\item{sample}{is a data frame return by \code{\link{read.LAMBDA850}} or
\code{\link{read.ULTRAPATH}}. It contains two  2 columns for wavelength (wl)
and Optical Depth (OD).}

\item{ID}{is the sample ID.}

\item{Station}{is the station code (optional).}

\item{Depth}{is the depth of the sample in meters (optional).}

\item{pathlength}{is the pathlength of the cuvette in meters.}

\item{DilutionFactor}{Is a multiplication factor to adjust the final Ag value if dilution was
performed in the lab (default=1).}
}
\value{
A list with all the computed parameters decribed above is return.
}
\description{
Spectrophotometer measures opttical density (depth) (OD) or Absorbance (A),
 which are unit less. This program converts OD into absorption coefficent
 in per meter. Then a series of spectral analyses is performed.
}
\details{
The processing steps includes:

1) Ag = 2.303*OD/pathlenght.

2) Apply a null correction assuming Ag(690+/-5) = 0.
  If a(500) is negative after the applicatioc of the NULL correction, then
  the minimum a values between 500 and 690 is assume for the null correction.
  So Ag.offset = Ag - NULL.POINT

3) Fit an exponential model to Ag.offset for the 350 to 500 nm range:
  Ag.fitted = Ag(440) exp(-S*(440-lambda)) + K.
  This is done using \code{\link{cdom_fit_exponential}}.

4) Compute spectral slope for the 275-295 and 350-400 ranges.
  Compute slope ratio using  \code{\link{cdom_slope_ratio}}.

5) Compute spectral curvature using \code{\link{cdom_spectral_curve}}.
}
\seealso{
The \pkg{cdom} package for more details on the CDOM spectral analysis.
}
\author{
Simon Belanger
}
