% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.Ap.R
\name{process.Ap}
\alias{process.Ap}
\title{Process raw optical depth to Ap spectra using filter-pad technique}
\usage{
process.Ap(sample, blank, Farea = NA, Vol.ml = NA, SpecType = NA, ID,
  Station, Repl, Depth)
}
\arguments{
\item{sample}{is a data frame return by \code{\link{read.LAMBDA850}} for the measured optical depth
of the filter containing the particles. It contains two  2 columns for wavelength (wl)
and Optical Depth (OD).}

\item{blank}{is a data frame return by \code{\link{read.LAMBDA850}} for the measured optical depth
of a blank filter. It contains two  2 columns for wavelength (wl)
and Optical Depth (OD).}

\item{Farea}{is the filter clearance area of particles in m^2. This field is mandatory.}

\item{Vol.ml}{is the volume filtered in ml. This field is mandatory.}

\item{SpecType}{is a string indicating either "Ap" for the total particle  or "Anap" for non-algal particles.
This field is mandatory.}

\item{ID}{is the sample ID.}

\item{Station}{is the station code.}

\item{Repl}{is a Capital Letter or number corresponding of the replicat ("A", "B",... or "1", etc)}

\item{Depth}{is the depth of the sample in meters.}
}
\value{
Returns a list containing metadata provided to the function as well as the vector of OD corrected for blank,
the OD of the blank filter and 3 estimation of particle absorption:
\itemize{
   \item{Ap_Stramski folowwing Stramski et al, App. Optics 2015}
   \item{Ap_RG following Rottgers and Genhke, App. Opt. (2012)}
   \item{Ap_4.5 using a constant beta factor of 4.5 (see Rottgers and Genhke, 2012)}
}
}
\description{
Spectrophotometer measures optical density (depth) (OD) or Absorbance (A),
 which are unit less. This program converts OD into absorption coefficent
 in per meter. Then a series of spectral analyses is performed.
}
