% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ULTRAPATH.R
\name{read.ULTRAPATH}
\alias{read.ULTRAPATH}
\title{Read CSV file from Ultrapath spectrophotometer}
\usage{
read.ULTRAPATH(filen)
}
\arguments{
\item{filen}{is the file name.}
}
\value{
It returns a data frame with 2 columns for wavelength (wl)
and Optical Depth (OD).
}
\description{
Read text file from the Ultrapath spectrophotometer.
}
\details{
The ultrapath data files have a header containing wavelength
specifications and than 16 optical depth spectra. The function extract the header information
and average the 16 spectra.
}
\author{
Simon Bélanger
}
