% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.compute.Aph.batch.R
\name{run.compute.Aph.batch}
\alias{run.compute.Aph.batch}
\title{Compute phytoplankton absorption spectrum in batch mode}
\usage{
run.compute.Aph.batch(
  log.file = "Ap_log_TEMPLATE.dat",
  data.path = "./",
  PLOT = FALSE
)
}
\arguments{
\item{log.file}{is the name of the ASCII file containing the list of
samples to process.}

\item{data.path}{is the path where the csv folder is located.
So the files must be stored in data.path/csv/.  Default is "./".}

\item{PLOT}{is a logical parameter indicating whether the plot are generated
in the RStudio interface. Default is FALSE.}
}
\value{
It produces plots with the spectra returned by \code{\link{compute.Aph}}
 and save the data in a RData file.
}
\description{
This is the third step of the Ap/Aph processing. It calls the \code{\link{compute.Aph}}
for each samples found in the Log file provided as input.
}
\seealso{
\code{\link{compute.Aph}}
}
\author{
Simon Belanger
}
