% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.process.Ag.batch.R
\name{run.process.Ag.batch}
\alias{run.process.Ag.batch}
\title{Process CDOM spectra in batch mode}
\usage{
run.process.Ag.batch(log.file = "Ag_log_TEMPLATE.dat", suffix = "",
  data.path = "./", instrument = "LAMBDA850")
}
\arguments{
\item{log.file}{is the name of the ASCII file containing the
list of samples to process (see details below).}

\item{suffix}{is an optional character string for file names that
have a suffix after the ID in the file name (e.g. "_01_CDOM").
Default is "" for no suffix. (see details below).}

\item{data.path}{is the path where the csv folder is located (txt folder in the case of ULTRAPATH).
So the files must be stored in data.path/csv/.  Default is "./".}

\item{instrument}{is the type of instrument, i.e. "LAMBDA850" or "ULTRAPATH".
When ULTRAPATH is set, two files are red for each sample. inclusing the reference sample,
which is an artificial sea water solution.
The default is LAMBDA850.}
}
\description{
This high level function reads an ASCII file containing the
list of samples to process. It calls \code{\link{process.Ag}}.
}
\details{
The most important thing to do before runing this programm is to prepare
the log.file. This file contains 6 fields : ID  Station  Depth	pathlength Ag.good DilutionFactor

ID is the sample ID. It is usually the base name of the CSV file.
   for example, 407839 ID will have the following file name: 407839.Sample.Raw.csv.
   where the ".Sample.Raw.csv" was automatically added by the Lambda850 software.
   NOTE: Sometime the analyst added a suffix between the ID and ".Sample.Raw.csv".
   For example, "_Ag" may have been added yielding a file name like 407839_Ag.Sample.Raw.csv.
   In these case you have to set parameter suffix="_Ag"

Station is the Station name. For example: "IML4", "L3_18", etc.

Depth is the depth of the sample in meters.

pathlength is the pathlength of the cuvette in meters (e.g. 0.1 when using 10-cm cuvette).

Ag.good is a binary field where 1 will process the sample while 0 will skip the sample.

DilutionFactor Is a  factor to adjust the final Ag value if dilution was
performed in the lab (default=1).

The program will automatically create two new folders in the data.path to store the results.
For each process sample, a png and a RData file is produce and stored in
data.path/png and data.path/RData, respectively.
}
\seealso{
\code{\link{process.Ag}}
}
\author{
Simon Belanger
}
