% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.process.Ap.batch.R
\name{run.process.Ap.batch}
\alias{run.process.Ap.batch}
\title{Process Particulate absorption spectra in batch mode}
\usage{
run.process.Ap.batch(log.file = "Ap_log_TEMPLATE.dat",
  data.path = "./")
}
\arguments{
\item{log.file}{is the name of the ASCII file containing the
list of samples to process (see details below).}

\item{data.path}{is the path where the csv folder is located (txt folder in the case of ULTRAPATH).
So the files must be stored in data.path/csv/.  Default is "./".}
}
\description{
This high level function reads an ASCII file containing the
list of samples to process. It calls \code{\link{process.Ap}}.
}
\details{
The most important thing to do before runing this programm is to prepare
the log.file. This file contains 11 fields (TAB delimated) :

  ID         Unique ID of the sample

  Repl       A letter corresponding to the replicat (A,B,C,etc)

  Station    Station name

  Date       Date of sampling

  Depth      Depth of the sample

  Vol        Filtered volume in mL

  Farea      clearance area of particles on filter in m^2

  blank.file ID if the reference blank filter

  Ap.good	   Boolean quality control indicator  (1=good ; 0 = not good)

  NAp.good   Boolean quality control indicator (1=good ; 0 = not good)

  process    Boolean (1=to be process ; 0 = to skip in the batch processing)

  NAP.method String indicating the method is retained to derive phytoplankton
  absorption ("Measured", "Fitted", "BS90_1"", "BS90_2")
}
